Login functionality and some small changes here and there
