+++ src/App.jsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import React, { useState } from "react";
 import reactLogo from "./assets/react.svg";
 import viteLogo from "/vite.svg";
 
@@ -14,17 +14,13 @@ import {
 } from "react-router-dom";
 
 import About from "./page/About.jsx";
-import Notes from "./page/SomeNotes.jsx";
+import Notes from "./page/Notes.jsx";
 import PageNotFound from "./page/PageNotFound.jsx";
 import Login from "./page/Login.jsx";
 import AddNote from "./page/AddNote.jsx";
 import AppLayout from "./layout/AppLayout.jsx";
-
 import CreateUser from "./features/CreateUser.jsx";
-
-
-import ProtectedRoutes from "./utils/ProtectedRoutes";
-
+import ProtectedRoutes from "./utils/ProtectedRoutes.jsx";
 
 function App() {
   const [isLoggedIn, setIsLoggedIn] = useState(false);
@@ -35,11 +31,9 @@ function App() {
   });
 
   return (
-
     <BrowserRouter>
       <Routes>
         <Route index element={<Navigate to="/login" />} />
-
         <Route
           path="/login"
           element={
@@ -51,6 +45,10 @@ function App() {
         />
         <Route path="/createUser" element={<CreateUser />} />
 
+        <Route path="*" element={<PageNotFound />} />
+
+        <Route path="/addNote" element={<AddNote />} />
+        
         <Route element={<ProtectedRoutes isLoggedIn={isLoggedIn} />}>
           <Route
             element={
@@ -62,14 +60,11 @@ function App() {
             }
           >
             <Route path="/notes" element={<Notes />} />
-
-            <Route path="*" element={<PageNotFound />} />
+            <Route path="/about" element={<About />} />
+            
+            
           </Route>
         </Route>
-
-        <Route path="/about" element={<About />} />
-
-        <Route path="/addNote" element={<AddNote />} />
       </Routes>
     </BrowserRouter>
   );
