+++ src/layout/AppLayout.jsx
@@ -4,7 +4,7 @@ import { Outlet } from "react-router-dom";
 function AppLayout({setIsLoggedIn, loggedInUser, setLoggedInUser}) {
     return (
       <>
-        <Header etIsLoggedIn={setIsLoggedIn} loggedInUser={loggedInUser} setLoggedInUser={setLoggedInUser}/>
+        <Header setIsLoggedIn={setIsLoggedIn} loggedInUser={loggedInUser} setLoggedInUser={setLoggedInUser}/>
         <Outlet />
       </>
     );
