--- src/page/CreateUser.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function CreateUser() {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default CreateUser
