+++ src/utils/globalVariables.js
@@ -1,0 +1,2 @@
+export const BASE_URL_DEV = "http://localhost:7070/api";
+export const BASE_URL_PROD = "https://api.tobiasrossen.com/api";
